fixed provisioning issue with if-then locks

+++ scripts/common.sh
@@ -7,29 +7,32 @@ set -euxo pipefail
 # Variable Declaration
 
 # DNS Setting
-#if [ ! -d /etc/systemd/resolved.conf.d ]; then
-#	sudo mkdir /etc/systemd/resolved.conf.d/
-#fi
-#cat <<EOF | sudo tee /etc/systemd/resolved.conf.d/dns_servers.conf
-#[Resolve]
-#DNS=${DNS_SERVERS}
-#EOF
-
-#sudo systemctl restart systemd-resolved
+if compgen -abc | grep systemd-resolved ;
+    then 
+        if [ ! -d /etc/systemd/resolved.conf.d ]; then
+            sudo mkdir /etc/systemd/resolved.conf.d/
+        fi
+    cat <<EOF | sudo tee /etc/systemd/resolved.conf.d/dns_servers.conf
+    [Resolve]
+    DNS=${DNS_SERVERS}
+EOF
+    sudo systemctl restart systemd-resolved
+fi
 
 # disable swap
 sudo swapoff -a
 
-# keeps the swaf off during reboot
+# keeps the swap off during reboot
 (crontab -l 2>/dev/null; echo "@reboot /sbin/swapoff -a") | crontab - || true
 sudo apt-get update -y
 
 # Create the .conf file to load the modules at bootup
+if [ ! -f /etc/modules-load.d/k8s.conf] ; then
 cat <<EOF | sudo tee /etc/modules-load.d/k8s.conf
 overlay
 br_netfilter
 EOF
-
+fi
 sudo modprobe overlay
 sudo modprobe br_netfilter
 
@@ -46,12 +49,18 @@ sudo sysctl --system
 ## Install CRIO Runtime
 
 sudo apt-get update -y
-apt-get install -y software-properties-common curl apt-transport-https ca-certificates
+apt-get install -y software-properties-common \
+curl \
+apt-transport-https \
+ca-certificates
 
-curl -fsSL https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/Release.key |
-    gpg --dearmor -o /etc/apt/keyrings/cri-o-apt-keyring.gpg
-echo "deb [signed-by=/etc/apt/keyrings/cri-o-apt-keyring.gpg] https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/ /" |
-    tee /etc/apt/sources.list.d/cri-o.list
+## Needs a conditional operator in order to perform redeploy properly
+if [ ! -f /etc/apt/keyrings/cri-o-apt-keyring.gpg ] ; then
+    curl -fsSL https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/Release.key |
+        gpg --dearmor -o /etc/apt/keyrings/cri-o-apt-keyring.gpg
+    echo "deb [signed-by=/etc/apt/keyrings/cri-o-apt-keyring.gpg] https://pkgs.k8s.io/addons:/cri-o:/prerelease:/main/deb/ /" |
+        tee /etc/apt/sources.list.d/cri-o.list
+fi
 
 sudo apt-get update -y
 sudo apt-get install -y cri-o
@@ -62,15 +71,14 @@ sudo systemctl start crio.service
 
 echo "CRI runtime installed successfully"
 
-sudo mkdir -p /etc/apt/keyrings
-curl -fsSL https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/Release.key | sudo gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
-echo "deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/ /" | sudo tee /etc/apt/sources.list.d/kubernetes.list
+if [ ! -f /etc/apt/sources.list.d/kubernetes.list ] ; then
+    sudo mkdir -p /etc/apt/keyrings
+    curl -fsSL https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/Release.key | sudo gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
+    echo "deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v$KUBERNETES_VERSION_SHORT/deb/ /" | sudo tee /etc/apt/sources.list.d/kubernetes.list
+fi
 
 sudo apt-get update -y
 sudo apt-get install -y kubelet="$KUBERNETES_VERSION" kubectl="$KUBERNETES_VERSION" kubeadm="$KUBERNETES_VERSION"
-sudo apt-get update -y
-sudo apt-get install -y jq
-
 sudo apt-get install -y \
 sudo \
 tmux \
@@ -80,15 +88,18 @@ nano \
 curl \
 wget \
 net-tools \
-which
+which \
+jq \
+build-essential \
+dkms \
+linux-headers-$(uname -r) 
 
 # Disable auto-update services
-sudo apt-mark hold kubelet kubectl kubeadm cri-o
-
-## Guest utilities
-#sudo apt install virtualbox-guest-utils
-sudo apt install -y build-essential dkms linux-headers-$(uname -r)
-
+sudo apt-mark hold \
+kubelet \
+kubectl \
+kubeadm \
+cri-o
 
 ## set localip address in kubelet_extra_args
 local_ip="$(ip --json a s | jq -r '.[] | if .ifname == "eth1" then .addr_info[] | if .family == "inet" then .local else empty end else empty end')"
